BotChooseBestFightWeapon: v6 should be float (matches v7) for fcom DWORD
FuzzyWeight returns double, but original DLL stored result into a float
local (truncating). With 'double v6', MSVC emits fld+fld+fcompp to compare
against float v7 (which it preloads). With 'float v6', the result is
fstp'd to memory and reloaded so the next compare can use 'fcom DWORD
PTR [v7]' directly — matching ref.

byte_diff 260 -> 0. MATCH 397 -> 398.

Source-commit: dc2c5d331897a94147e262b3068b691c589c6e8e

+++ botlib/botlib.c
@@ -31816,7 +31816,7 @@ void __cdecl BotChooseBestFightWeapon(bot_weaponstate_t *ws)
   weaponinfo_t *v2; // edi
   int v4; // ebx
   int v5; // eax
-  double v6; // st7
+  float v6; // st7
   float v7; // [esp+10h] [ebp-4h]
 
   v1 = (weaponconfig_t *)dword_10064080;
